Polyfill for ValidatorChain
Since the definitions for EventManager::attach differ between major versions, we
have to introduce a polyfill to alias version-specific implementations to the
ValidatorChain class.
